No longer enable async-bytecomp-package-mode
This was originally added while this library was still of Magit,
because we got many bug reports about seemingly impossible behavior
where we eventually figured out that some of Magit's libraries were
compiled with an older version already being loaded.  That lead to
outdated macro definitions and other things leak into the new byte
compiled files.

This is no longer necessary now that we move to GNU Elpa, because
unlike Melpa, it only publishes releases not snapshots.  So if we
do make in incompatible change, then we necessarily make a release
and update the dependency in Magit.